feat: Get Up Nav button to show up, and to show up as white
Apparently, the default color is black. So changing that to be white.
The Back/Up functionality seems to be hit or miss, as does the pinning
of the LinearLayout. This may have to do with the laying out of all the
diffrent pages of the ViewPager at the same time. Think about only doing
a couple. That hopefully would fix everything.